==> 03_SRA_figures_cod.R <==



################# Compare cod report
res1 <- readRDS("GoM_cod/SRA_cod_M02.rds")
res2 <- readRDS("GoM_cod/SRA_cod_MRAMP.rds")
res3 <- readRDS("GoM_cod/SRA_MC.rds")
res4 <- readRDS("GoM_cod/SRA_MCIM.rds")
res5 <- readRDS("GoM_cod/SRA_IM.rds")


compare_SRA(res1, res2, res3, res4, res5,
            filename = "report/GoM_cod/compare_cod", dir = getwd(), title = "ASAP3 GM cod comparison",
            s_name = c("NEFSCspring", "NEFSCfall", "MAspring"), 
            scenario = list(names = c("M02", "MRAMP", "MC", "MCIM", "IM")),
            open_file = FALSE)


no_panel_gap <- theme(panel.spacing = unit(0, "in"))
legend_bottom <- theme(legend.position = "bottom")


# Model peels
setup(3)
rr_OM <- sfLapply(list(res3, res5, res4), retrospective, nyr = 7, figure = FALSE)
rr_EM <- sfLapply(list(res1, res2), retrospective, nyr = 7, figure = FALSE)

model_name <- c("MC", "IM", "MCIM", "M02", "MRAMP")
rr_out <- lapply(1:5, function(i, x) x[[i]]@TS[, , 2] %>% reshape2::melt(value.name = "SSB") %>% mutate(Model = model_name[i]), 
                 x = c(rr_OM, rr_EM))
rr_out <- do.call(rbind, rr_out) %>% mutate(Model = factor(Model, levels = model_name))

rho_label <- data.frame(Model = model_name %>% factor(),
                        rho = sapply(c(rr_OM, rr_EM), function(x) summary(x)[2, 1] %>% format(trim = TRUE) %>% paste0("rho==", .)))


ggplot(rr_out %>% filter(Peel > 0) %>% mutate(Peel = factor(Peel)), aes(Year, SSB)) + 
  geom_line(aes(group = Peel, colour = Peel)) + facet_wrap(~ Model) + 
  geom_text(data = rho_label, x = -Inf, y = 0, vjust = "inward", hjust = -0.1, aes(label = rho), parse = TRUE) +
  geom_line(data = filter(rr_out, Peel == 0), size = 0.5, colour = "black") + 
  geom_text(data = data.frame(Model = model_name, label = paste0("(", letters[1:5], ")")),
            aes(label = label),
            inherit.aes = FALSE, x = -Inf, y = Inf, hjust = -0.1, vjust = 1.1
  ) + 
  coord_cartesian(ylim = c(0, 6e4)) + 
  scale_y_continuous(labels = scales::comma) + 
  gfplot::theme_pbs() + no_panel_gap + legend_bottom
ggsave("report/GoM_cod/cod_cond_rho.png", width = 6.5, height = 5)
