==> 99_Miscellaneous.R <==

library(dplyr)
library(ggplot2)

no_legend <- theme(legend.position = "none")
no_panel_gap <- theme(panel.spacing = unit(0, "in"))
legend_bottom <- theme(legend.position = "bottom")

# Figure 1 showing the retrospectives in halibut
halibut <- read.csv('halibut_retro.csv') %>% mutate(facet = ifelse(Model <= 1994, 1, 2))

TAC <- halibut %>% group_by(Model) %>% summarise(TAC = 0.1 * VB[which.max(Year)]) %>% 
  mutate(facet = ifelse(Model <= 1994, 1, 2), TAC2 = TAC/ifelse(facet == 1, 23.3, 33.6))
terminal <- halibut %>% group_by(Model) %>% summarise(VB = VB[which.max(Year)]) %>% 
  mutate(facet = ifelse(Model <= 1994, 1, 2), Year = Model)
  

a <- ggplot(halibut, aes(x = Year, y = VB, group = Model, colour = factor(Model))) + 
  facet_wrap(~facet, ncol = 1, scales = "free") + geom_line() + geom_point(data = terminal) + 
  geom_hline(yintercept = 0, col = "white") + 
  geom_vline(data = data.frame(facet = c(1, 2), yy = c(1996, 1995)), aes(xintercept = yy), col = "white") +
  ylab("Biomass estimate") + 
  geom_text(data = data.frame(x = Inf, y = Inf, facet = 1:2, txt = c("(a)", "(c)")), 
            vjust = 1.1, hjust = 1.1,
            inherit.aes = FALSE, aes(x = x, y = y, label = txt)) + 
  gfplot::theme_pbs() + no_legend + theme(strip.text.x = element_blank(), strip.text.y = element_blank())
b <- ggplot(TAC, aes(x = Model, y = TAC2)) + facet_wrap(~facet, ncol = 1, scales = "free") +
  geom_line() + geom_point(aes(colour = factor(Model))) + 
  geom_hline(yintercept = 0, col = "white") + 
  geom_text(data = data.frame(x = Inf, y = Inf, facet = 1:2, txt = c("(b)", "(d)")), 
            vjust = 1.1, hjust = 1.1,
            inherit.aes = FALSE, aes(x = x, y = y, label = txt)) + 
  labs(x = "Year", y = "Relative catch advice") + 
  coord_cartesian(ylim = c(0, 1.5)) + 
  gfplot::theme_pbs() + no_legend + theme(strip.text.x = element_blank(), strip.text.y = element_blank())

g <- cowplot::plot_grid(a, b, nrow = 1)
ggsave("report/intro_figure.png", height = 3, width = 5)
ggsave("report/intro_figure.png", g, height = 80, width = 170, dpi = 600, units = "mm")


###### Combine SSB and recruitment plot with both cod and pollock
no_panel_gap <- theme(panel.spacing = unit(0, "in"))
legend_bottom <- theme(legend.position = "bottom")


cod <- local({
  
  res1 <- readRDS("GoM_cod/SRA_cod_M02.rds")
  res2 <- readRDS("GoM_cod/SRA_cod_MRAMP.rds")
  res3 <- readRDS("GoM_cod/SRA_MC.rds")
  res4 <- readRDS("GoM_cod/SRA_MCIM.rds")
  res5 <- readRDS("GoM_cod/SRA_IM.rds")
  
  Model <- data.frame(name = c("MC", "IM", "MCIM", "M02", "MRAMP"), type = c("OM", "OM", "OM", "AM", "AM"))
  SSB <- lapply(list(res3, res5, res4, res1, res2), function(x) x@SSB[1, ]) %>% do.call(cbind, .) %>% 
    structure(dimnames = list(Year = 1982:2019, name = Model$name)) %>% reshape2::melt() %>% mutate(y = "SSB") %>%
    left_join(Model, by = "name")
  
  RR <- lapply(list(res3, res5, res4, res1, res2), function(x) x@Misc[[1]]$R) %>% do.call(cbind, .) %>% 
    structure(dimnames = list(Year = 1982:2019, name = Model$name)) %>% reshape2::melt() %>% mutate(y = "Recruitment") %>%
    left_join(Model, by = "name")
  
  rbind(SSB, RR) %>% 
    mutate(y = factor(y, levels = c("SSB", "Recruitment")), 
           name = factor(name, levels = Model$name), Stock = "Cod") %>% 
    filter(value <= 1e6) %>%
    ggplot(aes(x = Year, y = value, linetype = name, colour = name)) + geom_line() +
    geom_hline(yintercept = 0, colour = "white") +
    facet_grid(y ~ Stock, scales = "free_y", switch = "y") + ylab(NULL) + 
    geom_text(data = data.frame(x = Inf, yy = Inf, y = c("SSB", "Recruitment"), txt = c("(a)", "(b)")),
              aes(x, yy, label = txt), hjust = 1.1, vjust = 1.1,
              inherit.aes = FALSE) + 
    scale_y_continuous(labels = scales::comma, n.breaks = 5) +
    scale_linetype_manual(values = c("MC" = 1, "IM" = 1, "MCIM" = 1, "M02" = 2, "MRAMP" = 2)) +
    gfplot::theme_pbs() + no_panel_gap + legend_bottom + 
    theme(legend.title = element_blank(), strip.placement = "outside") +
    guides(linetype = guide_legend(nrow = 2), colour = guide_legend(nrow = 2))
})


pollock <- local({
  
  res1 <- readRDS("pollock/SRA_pollock_base.rds")
  res2 <- readRDS("pollock/SRA_pollock_flatsel.rds")
  res3 <- readRDS("pollock/SRA_SS.rds")
  res4 <- readRDS("pollock/SRA_SWB.rds")
  res5 <- readRDS("pollock/SRA_SWF.rds")
  
  # SSB and R
  Model <- data.frame(name = c("SS", "SWB", "SWF", "Base", "FlatSel"), type = c("OM", "OM", "OM", "AM", "AM"))
  SSB <- lapply(list(res3, res4, res5, res1, res2), function(x) x@SSB[1, ]) %>% do.call(cbind, .) %>% 
    structure(dimnames = list(Year = 1970:2019, name = Model$name)) %>% reshape2::melt() %>% mutate(y = "SSB") %>%
    left_join(Model, by = "name")
  
  RR <- lapply(list(res3, res4, res5, res1, res2), function(x) x@Misc[[1]]$R) %>% do.call(cbind, .) %>% 
    structure(dimnames = list(Year = 1970:2019, name = Model$name)) %>% reshape2::melt() %>% mutate(y = "Recruitment") %>%
    left_join(Model, by = "name")
  
  rbind(SSB, RR) %>% 
    mutate(y = factor(y, levels = c("SSB", "Recruitment")), 
           name = factor(name, levels = Model$name), Stock = "Pollock") %>% 
    filter(value <= 1e6) %>%
    ggplot(aes(x = Year, y = value, linetype = name, colour = name)) + geom_line() +
    geom_hline(yintercept = 0, colour = "white") +
    geom_text(data = data.frame(x = Inf, yy = Inf, y = c("SSB", "Recruitment"), txt = c("(c)", "(d)")),
              aes(x, yy, label = txt), hjust = 1.1, vjust = 1.1,
              inherit.aes = FALSE) + 
    facet_grid(y ~ Stock, scales = "free_y", switch = "y") + ylab(NULL) +
    scale_y_continuous(labels = scales::comma, n.breaks = 5) +
    scale_linetype_manual(values = c("SS" = 1, "SWB" = 1, "SWF" = 1, "Base" = 2, "FlatSel" = 2)) +
    gfplot::theme_pbs() + no_panel_gap + legend_bottom + 
    theme(legend.title = element_blank(), strip.placement = "outside") +
    guides(linetype = guide_legend(nrow = 2), colour = guide_legend(nrow = 2))
})


g <- ggpubr::ggarrange(cod, pollock, ncol = 2)
ggsave("report/model_compare.png", height = 4, width = 8)
ggsave("report/model_compare.png", g, height = 125, width = 170, dpi = 600, units = "mm")
